More zmodem cleanups.
* Start to use types from cstdint vs. built in types.
* Make zmodemlog use fmt::format format types.
* Cleanup the logging a bit more
* Implement ZIFlush to purge the inbound queueue.  We can't purge
  outbound since we never have anything waiting to be sent.

+++ bbs/prot/reference/checkcrc.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #ifndef	lint
 static const char rcsid[] = "$Id$" ;
 #endif
@@ -98,9 +99,9 @@ static	int	addcrc(int data, int crc) ;
 main()
 {
 	u_short	crc16 ;
-	u_long	crc32 ;
+	uint32_t	crc32 ;
 	int	i,j ;
-	u_long	k ;
+	uint32_t	k ;
 	u_char	*msg = msg0 ;
 	int	len = sizeof(msg0) ;
